+++ src/main/kotlin/com/macoder/styling/authentication/persistence/MemberRepository.kt
@@ -5,7 +5,7 @@ import com.macoder.styling.common.entity.Member
 
 import org.springframework.data.jpa.repository.JpaRepository
 
-interface MemberRepository : JpaRepository<Member, Int> {
+interface MemberRepository : JpaRepository<Member, Long> {
     fun findByAccount(account: String): Member?
     fun findAllByType(type: MemberType): List<Member>
 }
